Update of Practice page to use Suspense wrapper
